creating list_rentals_id class + list_people class + adding instance variable to library class
